backend/requirements.txt
fastapi==0.104.0
uvicorn==0.23.2
python-multipart==0.0.6
python-dotenv==1.0.0
openai==0.28.0
pydantic==2.4.2
requests==2.31.0
opencv-python==4.8.1.78
numpy==1.26.0
# Optional dependencies (will be skipped if installation fails)
groq
ultralytics 